moves validation error
