+++ lib/bureaucrat/fields/float_field.rb
@@ -1,4 +1,5 @@
 require 'bureaucrat/fields/integer_field'
+require 'bureaucrat/validation_error'
 
 module Bureaucrat
   module Fields
