+++ lib/bureaucrat/fields/range_field.rb
@@ -1,4 +1,5 @@
 require 'bureaucrat/fields/field'
+require 'bureaucrat/validation_error'
 require 'bureaucrat/widgets/range'
 
 module Bureaucrat
